refactor python run all
